exwebbekend/backend/requirements.txt
fastapi==0.114.0
uvicorn[standard]==0.30.6
SQLAlchemy[asyncio]==2.0.35
asyncpg==0.29.0
pydantic==2.8.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
